feat: created root file for types
