+++ src/models/orders.store.ts
@@ -1,19 +1,9 @@
+import { Order, OrderModel } from "@/types";
 import { Instance, types } from "mobx-state-tree";
-
-const Order = types.model({
-  _id: types.string,
-  teamMember: types.optional(types.array(types.string), []),
-  status: types.optional(types.string, ""),
-  date: types.string,
-  totalPrice: types.optional(types.string, ""),
-  products: types.optional(types.array(types.string), []),
-});
-
-export type Order = Instance<typeof Order>;
 
 export const OrderStore = types
   .model({
-    orders: types.array(Order),
+    orders: types.array(OrderModel),
     orderSelected: types.optional(types.string, ""),
   })
   .views((store) => ({
@@ -25,7 +15,7 @@ export const OrderStore = types
     setOrders(orders) {
       store.orders = orders;
     },
-    setSelectedOrder(orderId) {
+    setSelectedOrder(orderId: string) {
       store.orderSelected = orderId;
     },
   }));
