+++ src/models/teams.store.ts
@@ -1,13 +1,5 @@
-import { Instance, types } from "mobx-state-tree";
-import { TeamMemberModel } from "./member.store";
-
-export const TeamModel = types.model({
-  _id: types.string,
-  name: types.optional(types.string, ""),
-  teamMembers: types.optional(types.array(TeamMemberModel), []),
-});
-
-export type Team = Instance<typeof TeamModel>
+import { types } from "mobx-state-tree";
+import { TeamModel, Team } from "@/types";
 
 export const TeamStore = types
   .model({
@@ -18,7 +10,7 @@ export const TeamStore = types
       store.teams = teams;
     },
 
-    addTeam(team) {
+    addTeam(team: Team) {
       store.teams.push(team);
     },
   }));
