+++ src/types/product.ts
@@ -1,0 +1,27 @@
+import { types, Instance } from "mobx-state-tree";
+
+export const PRODUCT_STATUSES = ["Available", "Delivered"] as const;
+
+export type ProductStatus = (typeof PRODUCT_STATUSES)[number];
+
+export const Product = types.model({
+  _id: types.string,
+  name: types.optional(types.string, ""),
+  description: types.optional(types.string, ""),
+  category: types.optional(types.string, ""),
+  model: types.optional(types.string, ""),
+  color: types.optional(types.string, ""),
+  screen: types.optional(types.string, ""),
+  keyboard: types.optional(types.string, ""),
+  processor: types.optional(types.string, ""),
+  ram: types.optional(types.string, ""),
+  storage: types.optional(types.string, ""),
+  gpu: types.optional(types.string, ""),
+  serialNumber: types.optional(types.string, ""),
+  price: types.optional(types.string, ""),
+  status: types.optional(types.string, ""),
+  imgUrl: types.optional(types.string, ""),
+  quantity: types.number,
+});
+
+export type Product = Instance<typeof Product>;
